requirements.txt
# Interface e Visual
colorama>=0.4.4
pyfiglet>=0.8.0
rich>=13.0.0

# Validação e Dados
pydantic>=2.0.0
typer>=0.9.0

# Manipulação de Dados
pandas>=1.5.0
ruamel.yaml>=0.17.0

# Testes
pytest>=7.0.0
